Split RenderSprites out of RenderSimple.
RenderSprites handles sprite drawing independently
from the extra bits needed by actors that only use
sprites.
