+++ OpenRA.Game/Traits/Render/RenderSimple.cs
@@ -16,23 +16,14 @@ using OpenRA.FileFormats;
 
 namespace OpenRA.Traits
 {
-	public class RenderSimpleInfo : ITraitInfo, LocalCoordinatesModelInfo
+	public class RenderSimpleInfo : RenderSpritesInfo, LocalCoordinatesModelInfo
 	{
-		[Desc("Defaults to the actor name.")]
-		public readonly string Image = null;
-		[Desc("custom palette name")]
-		public readonly string Palette = null;
-		[Desc("custom PlayerColorPalette: BaseName")]
-		public readonly string PlayerPalette = "player";
-		[Desc("Change the sprite image size.")]
-		public readonly float Scale = 1f;
-
 		[Desc("Number of facings for gameplay calculations. -1 indiciates auto-detection from sequence")]
 		public readonly int QuantizedFacings = -1;
 
 		[Desc("Camera pitch the sprite was rendered with. Used to determine rotation ellipses")]
 		public readonly WAngle CameraPitch = WAngle.FromDegrees(40);
-		public virtual object Create(ActorInitializer init) { return new RenderSimple(init.self); }
+		public override object Create(ActorInitializer init) { return new RenderSimple(init.self); }
 
 		public virtual IEnumerable<IRenderable> RenderPreview(ActorInfo ai, PaletteReference pr)
 		{
@@ -43,73 +34,21 @@ namespace OpenRA.Traits
 		}
 	}
 
-	public class RenderSimple : IRender, ILocalCoordinatesModel, IAutoSelectionSize, ITick, INotifyOwnerChanged
+	public class RenderSimple : RenderSprites, ILocalCoordinatesModel, IAutoSelectionSize
 	{
-		public Dictionary<string, AnimationWithOffset> anims = new Dictionary<string, AnimationWithOffset>();
-
-		public static Func<int> MakeFacingFunc(Actor self)
-		{
-			var facing = self.TraitOrDefault<IFacing>();
-			if (facing == null) return () => 0;
-			return () => facing.Facing;
-		}
-
-		public Animation anim
-		{
-			get { return anims[""].Animation; }
-			protected set { anims[""] = new AnimationWithOffset(value,
-				anims[""].OffsetFunc, anims[""].DisableFunc, anims[""].ZOffset); }
-		}
-
-		public static string GetImage(ActorInfo actor)
-		{
-			var Info = actor.Traits.Get<RenderSimpleInfo>();
-			return Info.Image ?? actor.Name;
-		}
-
-		public string GetImage(Actor self)
-		{
-			if (cachedImage != null)
-				return cachedImage;
-
-			return cachedImage = GetImage(self.Info);
-		}
-
 		RenderSimpleInfo Info;
-		string cachedImage = null;
-		bool initializePalette = true;
-		protected PaletteReference palette;
 
 		public RenderSimple(Actor self, Func<int> baseFacing)
+			: base(self, baseFacing)
 		{
 			anims.Add("", new Animation(GetImage(self), baseFacing));
 			Info = self.Info.Traits.Get<RenderSimpleInfo>();
 		}
 
-		public RenderSimple(Actor self) : this( self, MakeFacingFunc(self) )
+		public RenderSimple(Actor self)
+			: this(self, MakeFacingFunc(self))
 		{
 			anim.PlayRepeating("idle");
-		}
-
-		protected virtual string PaletteName(Actor self)
-		{
-			return Info.Palette ?? Info.PlayerPalette + self.Owner.InternalName;
-		}
-
-		protected void UpdatePalette() { initializePalette = true; }
-		public void OnOwnerChanged(Actor self, Player oldOwner, Player newOwner) { UpdatePalette(); }
-
-		public virtual IEnumerable<IRenderable> Render(Actor self, WorldRenderer wr)
-		{
-			if (initializePalette)
-			{
-				palette = wr.Palette(PaletteName(self));
-				initializePalette = false;
-			}
-
-			foreach (var a in anims.Values)
-				if (a.DisableFunc == null || !a.DisableFunc())
-					yield return a.Image(self, wr, palette, Info.Scale);
 		}
 
 		public int2 SelectionSize(Actor self)
@@ -118,12 +57,6 @@ namespace OpenRA.Traits
 			                                && b.Animation.CurrentSequence != null)
 				.Select(a => (a.Animation.Image.size*Info.Scale).ToInt2())
 				.FirstOrDefault();
-		}
-
-		public virtual void Tick(Actor self)
-		{
-			foreach (var a in anims.Values)
-				a.Animation.Tick();
 		}
 
 		protected virtual string NormalizeSequence(Actor self, string baseSequence)
